🎉 Initial commit
This commit initializes this repository with a README and default
license (CC BY-SA 4.0). The creation for this repository is motivated by
my effort to prioritize my own work while opening up my workstreams for
more collaboration by my Red Hat colleagues and members of various Red
Hat upstream communities. My vision is for this repository to become a
useful tool for Community Architects to engage with their communities
publicly, while also allowing more transparency into our workstreams to
each other. This way, we can see what each other is working on,
collaborate on shared projects (e.g. Red Hat Summit), and also mentor
others so the ways in which we work are better standardized.

This is the ambition. Before I introduce this more widely, first I want
to trial it for myself, just with Fedora, and see how successful I am.
If it works well, then I will seek more feedback and hopefully broaden
the pool of collaborators.

So, here is to that! Cheers.

Signed-off-by: Justin W. Flory (he/him) <jwf@redhat.com>