🔧 Issue templates: Add default template for Fedora
This commit introduces a new default issue template for Fedora. This
integrates the typical format that I use for issues that helps me frame
the context for a task. It also uses the Fedora issue label and assigns
me as the default assignee.

This prefaces the creation of new issues in this repository for Fedora.

Signed-off-by: Justin W. Flory (he/him) <jwf@redhat.com>